+++ gDRcore/R/utils.R
@@ -249,3 +249,15 @@ rbindList <- function(x) {
 rbindParallelList <- function(x, name) {
   S4Vectors::DataFrame(do.call(plyr::rbind.fill, lapply(x, function(x) as.data.frame("[[" (x, name)))))
 }
+
+#' Function for detection of available cores used in parallel computing
+#'
+#' @return number of available cores
+#' @export
+detect_cores <- function() {
+  x <- as.numeric(Sys.getenv("NUM_CORES"))
+  if (is.na(x)) {
+    x <- parallel::detectCores() - 1
+  }
+  x
+}