+++ gDRcore/inst/CHANGELOG.md
@@ -2,9 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
-#### [1.1.22] - 2022-01-25
+<<<<<<< HEAD
+#### [1.1.24] - 2022-01-25
 #### Update
 - standardize/improve CI
+
+#### [1.1.23] - 2022-01-24
+#### Update
+- switch from `cores` variable to `detect_cores()` function
+
+#### [1.1.22] - 2022-01-24
+#### Update
+- fix wrong type of `NUM_CORES` env variable
 
 #### [1.1.21] - 2022-01-03
 #### Update